refactor: rethink app architecture
Give the connector registry a home in the domain layer.

`ConnectorRegistry`, its default value, and its lookup/composition helpers
lived in `lib/config/ConfigProvider.tsx` — a React client component — so 12
domain modules imported their central type out of the UI layer, and two
import cycles existed:

  lib/config/connectorUtils.ts  <-> lib/config/ConfigProvider.tsx
  lib/connectors/.../animated-image-chain.ts <-> lib/config/ConfigProvider.tsx

All of it moves to `lib/connectors/registry.ts`. `ConfigProvider` keeps only
what it is for: React state and project-scoped plugin composition.

Also collapses the repeated "resolve the default provider, then construct the
connector" idiom into `createDefaultConnector(registry, type, plugins?)`.
