+++ app/components/canvas/__tests__/useGenerateAll.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi, beforeEach } from "vitest";
 import type { Descendant } from "slate";
-import type { ConnectorRegistry } from "@/lib/config/ConfigProvider";
+import type { ConnectorRegistry } from "@/lib/connectors/registry";
 import type { GenerationJob } from "@/lib/generation/queue";
 import type { CanvasContentElement, SceneElement } from "@/lib/canvas/types";
