+++ app/components/canvas/hooks/useProjectRehydrate.test.ts
@@ -3,7 +3,7 @@ import { withHistory } from "slate-history";
 import { describe, expect, it } from "vitest";
 import { serializeOSMLWithScenes } from "@/lib/canvas/osmlSerializer";
 import { SCENE_TYPE, type SceneElement } from "@/lib/canvas/types";
-import type { ConnectorRegistry } from "@/lib/config/ConfigProvider";
+import type { ConnectorRegistry } from "@/lib/connectors/registry";
 import { rehydrateProjectEditor } from "./useProjectRehydrate";
 
 const connector = {
