+++ lib/connectors/animated_image/plugins/__tests__/animated-image-chain.test.ts
@@ -6,7 +6,7 @@ vi.mock("@/lib/connectors/factory", () => ({
 	createConnector: () => ({ generate: generateMock }),
 }));
 
-import type { ConnectorRegistry } from "@/lib/config/ConfigProvider";
+import type { ConnectorRegistry } from "@/lib/connectors/registry";
 import type {
 	AnimatedImageGenerateParams,
 	AssetResult,
